chore(network): move protobuf conversion into block header module
